Wl-66 SAK-11636 Implement caching of refershUser.
This means that we refresh users at most every 5 minutes, this prevents services such as WebDAV causing big problems for the groupprovider (LDAP).
We add a dependency on ehcache as it's in shared and we need a custom configured cache. Adding a MemCacheFactory didn't seem easy (no nice place to put it).